+++ packages/gallery/src/components/helpers/cssLayoutsHelper.js
@@ -5,12 +5,50 @@ import { cssScrollHelper } from './cssScrollHelper.js';
 const desktopWidths = [480, 768, 1024, 1280, 1440, 1680, 1920, 2560];
 const mobileWidths = [320]; //, 375, 414, 480, 600, 768, 900]; (mobile is currently fixed to 320px)
 
-const getImageStyle = item => ({
+const getImageStyle = (item, styleParams) => ({
   top: item.offset.top,
   left: item.offset.left,
   width: item.width,
-  height: item.height,
+  height: item.height + (styleParams.externalInfoHeight || 0),
 });
+
+const createExactCssForItems = (galleryItems, styleParams) => {
+  let cssStr = '';
+  galleryItems.forEach((item, i) => {
+    const id = cssScrollHelper.getDomId(item);
+    const style = getImageStyle(item, styleParams);
+    const T = `top:${style.top}px;`;
+    const L = `left:${style.left}px;`;
+    const W = `width:${style.width}px;`;
+    const H = `height:${style.height}px;`;
+    cssStr += `#${id} {${T}${L}${W}${H}}`;
+    // cssStr += `#${id} .gallery-item-wrapper, #${id} .gallery-item-hover, #${id} .gallery-item {${Wvw}${Hvw}}`;
+  });
+  return cssStr;
+};
+
+const createCssFromLayout = (layout, styleParams, width) => {
+  let cssStr = '';
+  const layoutWidth = width - styleParams.imageMargin * 2;
+  const getRelativeDimension = val =>
+    Math.round(10000 * (val / layoutWidth)) / 100;
+  layout.items.forEach((item, i) => {
+    const id = cssScrollHelper.getDomId(item);
+    if (i < 50) {
+      const style = getImageStyle(item, styleParams);
+      const Tvw = `top:${getRelativeDimension(style.top)}vw;`;
+      const Wvw = `width:${getRelativeDimension(style.width)}vw;`;
+      const Hvw = `height:${getRelativeDimension(style.height)}vw;`;
+      const Lpc = `left:${getRelativeDimension(style.left)}%;`;
+      const Wpc = `width:${getRelativeDimension(style.width)}%;`;
+      cssStr += `#${id} {${Tvw}${Lpc}${Wpc}${Hvw}}`;
+      cssStr += `#${id} .gallery-item-wrapper, #${id} .gallery-item-hover, #${id} .gallery-item {${Wvw}${Hvw}}`;
+    } else {
+      cssStr += `#${id}{display:none;}`;
+    }
+  });
+  return cssStr;
+};
 
 const createCssFromLayouts = (layouts, styleParams, widths) => {
   const cssStrs = [];
@@ -24,29 +62,7 @@ const createCssFromLayouts = (layouts, styleParams, widths) => {
         ? ''
         : `@media only screen and (min-width: ${(lastWidth * 2 + width) /
             3}px) {`;
-      const layoutWidth = width - styleParams.imageMargin * 2;
-      const getRelativeDimension = val =>
-        Math.round(10000 * (val / layoutWidth)) / 100;
-      layout.items.forEach((item, i) => {
-        const id = cssScrollHelper.getDomId(item);
-        if (i < 50) {
-          const style = getImageStyle(item);
-          const Tvw = `top:${getRelativeDimension(style.top)}vw !important;`;
-          const Wvw = `width:${getRelativeDimension(
-            style.width,
-          )}vw !important;`;
-          const Hvw = `height:${getRelativeDimension(
-            style.height,
-          )}vw !important;`;
-          const Lpc = `left:${getRelativeDimension(style.left)}% !important;`;
-          const Wpc = `width:${getRelativeDimension(style.width)}% !important;`;
-          cssStr += `#${id} {${Tvw}${Lpc}${Wpc}${Hvw}}`;
-          cssStr += `#${id} .gallery-item-wrapper, #${id} .gallery-item-hover, #${id} .gallery-item {${Wvw}${Hvw}}`;
-        } else {
-          cssStr += `#${id}{display:none;}`;
-        }
-      });
-
+      cssStr += createCssFromLayout(layout, styleParams, width);
       cssStr += isFirstMediaQuery ? '' : `}`;
       cssStrs.push(cssStr);
     }
@@ -55,16 +71,36 @@ const createCssFromLayouts = (layouts, styleParams, widths) => {
   return cssStrs;
 };
 
-export const createCssLayouts = (layoutParams, isMobile) => {
-  const widths = isMobile ? mobileWidths : desktopWidths;
-  const cssLayouts = widths.map(width => {
-    const _layoutParams = {
-      ...layoutParams,
-      ...{
-        container: { ...layoutParams.container, galleryWidth: width, width },
-      },
-    };
-    return createLayout(_layoutParams);
-  });
-  return createCssFromLayouts(cssLayouts, layoutParams.styleParams, widths);
+export const createCssLayouts = ({
+  isApproximation,
+  galleryItems,
+  layoutParams,
+  isMobile,
+}) => {
+  if (isApproximation) {
+    const widths = isMobile ? mobileWidths : desktopWidths;
+    const cssLayouts = widths.map(width => {
+      const _layoutParams = {
+        ...layoutParams,
+        ...{
+          container: { ...layoutParams.container, galleryWidth: width, width },
+        },
+      };
+      return createLayout(_layoutParams);
+    });
+    return createCssFromLayouts(cssLayouts, layoutParams.styleParams, widths);
+  } else {
+    // const chunkSize = 10;
+    // const itemsBatchs = [];
+    // for (let i = 0; i < galleryItems.length; i += chunkSize) {
+    //   itemsBatchs.push(galleryItems.slice(i, i + chunkSize));
+    // }
+    // return galleryItems.map(items =>
+    //   createExactCssForItems(items, layoutParams.styleParams)
+    // );
+    const exactCss = [];
+    exactCss.push(createExactCssForItems(galleryItems, layoutParams.styleParams));
+    return exactCss;
+
+  }
 };