+++ packages/gallery/src/components/item/itemView.js
@@ -753,26 +753,30 @@ class ItemView extends GalleryComponent {
   }
 
   getItemContainerStyles() {
-    const { styleParams, style, transform } = this.props;
+    const { styleParams, style, transform, offset } = this.props;
     const containerStyleByStyleParams = getContainerStyle(styleParams);
-
-    if (utils.isSSR()) {
-      return { ...transform, ...containerStyleByStyleParams }; //in SSR, item styles arrive from css. htydrate phase should change these
-    }
 
     const itemStyles = {
       overflowY: styleParams.isSlideshow ? 'visible' : 'hidden',
       position: 'absolute',
-      top: this.props.offset.top,
-      left: this.props.offset.left,
       right: 'auto',
       bottom: 'auto',
-      width: style.width,
-      height: style.height + (styleParams.externalInfoHeight || 0),
       margin: styleParams.oneRow ? styleParams.imageMargin + 'px' : 0,
     };
 
+    if (utils.isSSR()) {
+      return { ...itemStyles, ...transform, ...containerStyleByStyleParams }; //in SSR, item styles arrive from css. htydrate phase should change these
+    };
+
+    const positionStyles = {
+      top: offset.top,
+      left: offset.left,
+      width: style.width,
+      height: style.height + (styleParams.externalInfoHeight || 0),
+    };
+
     const styles = {
+      ...positionStyles,
       ...itemStyles,
       ...transform,
       ...containerStyleByStyleParams,